Implemented Model View Presenter, while removing multiple images for now
