+++ app/src/main/java/org/wit/placemark/activities/HillfortAdapter.kt
@@ -40,8 +40,8 @@ class HillfortAdapter constructor(
     fun bind(hillfort: HillfortModel, listener: HillfortListener) {
       itemView.hillfortTitle.text = hillfort.title
       itemView.description.text = hillfort.description
-      if(hillfort.image.size>0)
-        itemView.imageIcon.setImageBitmap(readImageFromPath(itemView.context, hillfort.image[0]))
+//      if(hillfort.image.size>0)
+      itemView.imageIcon.setImageBitmap(readImageFromPath(itemView.context, hillfort.image))
       itemView.setOnClickListener { listener.onHillfortClick(hillfort) }
     }
   }
