+++ app/src/main/java/org/wit/placemark/activities/HillfortPresenter.kt
@@ -1,0 +1,76 @@
+package org.wit.placemark.activities
+
+import android.content.Intent
+import org.jetbrains.anko.intentFor
+import org.wit.placemark.helpers.showImagePicker
+import org.wit.placemark.main.MainApp
+import org.wit.placemark.models.HillfortModel
+import org.wit.placemark.models.Location
+
+class HillfortPresenter(val view: HillfortView) {
+
+    val IMAGE_REQUEST = 1
+    val LOCATION_REQUEST = 2
+
+    var hillfort = HillfortModel()
+    var location = Location(52.245696, -7.139102, 15f)
+    var app: MainApp
+    var edit = false;
+
+    init {
+        app = view.application as MainApp
+        if (view.intent.hasExtra("hillfort_edit")) {
+            edit = true
+            hillfort = view.intent.extras?.getParcelable<HillfortModel>("hillfort_edit")!!
+            view.showHillfort(hillfort)
+        }
+    }
+
+    fun doAddOrSave(title: String, description: String) {
+        hillfort.title = title
+        hillfort.description = description
+        if (edit) {
+            app.hillforts.update(hillfort)
+        } else {
+            app.hillforts.create(hillfort)
+        }
+        view.finish()
+    }
+
+    fun doCancel() {
+        view.finish()
+    }
+
+    fun doDelete() {
+        app.hillforts.delete(hillfort)
+        view.finish()
+    }
+
+    fun doSelectImage() {
+        showImagePicker(view, IMAGE_REQUEST)
+    }
+
+    fun doSetLocation() {
+        if (hillfort.zoom != 0f) {
+            location.lat = hillfort.lat
+            location.lng = hillfort.lng
+            location.zoom = hillfort.zoom
+        }
+        view.startActivityForResult(view.intentFor<MapActivity>().putExtra("location", location), LOCATION_REQUEST)
+    }
+
+    fun doActivityResult(requestCode: Int, resultCode: Int, data: Intent) {
+        when (requestCode) {
+            IMAGE_REQUEST -> {
+                hillfort.image = data.data.toString()
+                view.showHillfort(hillfort)
+            }
+            LOCATION_REQUEST -> {
+                location = data.extras?.getParcelable<Location>("location")!!
+                hillfort.lat = location.lat
+                hillfort.lng = location.lng
+                hillfort.zoom = location.zoom
+            }
+        }
+    }
+}
