+++ app/src/main/java/org/wit/placemark/models/HillfortModel.kt
@@ -7,7 +7,8 @@ import kotlinx.android.parcel.Parcelize
 data class HillfortModel(var id: Long = 0,
                          var title: String = "",
                          var description: String = "",
-                         var image: MutableList<String> = mutableListOf(),
+                         var image: String = "",
+//                         MutableList<String> = mutableListOf(),
                          var lat : Double = 0.0,
                          var lng: Double = 0.0,
                          var zoom: Float = 0f) : Parcelable
