+++ app/src/main/res/layout-land/activity_hillfort.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    tools:context="org.wit.placemark.activities.HillfortActivity">
+    tools:context="org.wit.placemark.activities.HillfortView">
 
   <RelativeLayout
       android:layout_width="wrap_content"
