+++ app/src/main/res/layout/activity_hillfort.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.wit.placemark.activities.HillfortActivity">
+    tools:context="org.wit.placemark.activities.HillfortView">
 
   <RelativeLayout
       android:layout_width="match_parent"
